string: Move stpcpy and mempcpy symbol redirection to symbol-hacks.h
Commit 939da411433 added symbols redirections to handle ISO C
namespace, however some compiler does not support redeclare the
function prototype and moving these defintions to exported header
it not a good practice (it exposes a internal implementation and
it would require to add macros to define it only internally).

Instead this patch replaces the symbol redirections by direct asm
alias, as done to handle libcall generation done by compiler on
some loop optimizations.

Checked on most of affected ABIs.